% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{hms}
\alias{hms}
\alias{hms.POSIXlt}
\alias{hms.POSIXct}
\title{Hour-Minute-Second}
\usage{
hms(x)

\method{hms}{POSIXlt}(x)

\method{hms}{POSIXct}(x)
}
\arguments{
\item{x}{a \link[base]{POSIXlt} or \link[base]{POSIXct} object}
}
\value{
Function \code{\link[=hms]{hms()}} returns a \link[base]{difftime} object.
}
\description{
Function \link[readxl]{read_excel} might read a date-less time stamp
(e.g., \verb{3:00:00 PM} as \verb{1899-12-31 15:00:00 UTC}).
}
\seealso{
\link[base]{ISOdatetime}
}
\keyword{internal}
